Added loaders.py with classes to load input files and config.py with settings
