+++ commerce/bookstore/python/config.py
@@ -1,0 +1,2 @@
+data_path = "data/"  # path to csv files to import/load data
+db = "bookstore"  # Name of the DB to connect on the TypeDB
